+++ src/state.rs
@@ -65,4 +65,5 @@ impl State {
 pub enum Modal {
   AddProc { input: Input },
   RemoveProc { id: usize },
+  Quit,
 }